packages/k3splan/pyproject.toml
[project]
name = "k3splan"
version = "0.1.0"
description = "Declarative planning engine for k3sctl"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.11.0",
    "ruamel.yaml>=0.18.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

